Rename GeneralData to PickledData
