+++ aiida_workgraph/orm/__init__.py
@@ -1,3 +1,3 @@
-from .general_data import PickledData
+from .pickled_data import PickledData
 
 __all__ = ("PickledData",)
